refactor(app): move internal/skill → app/skill
SkillLoader, watcher, npx, validate, and config (plus their tests)
move into the app module. Config schema stays untouched —
internal/config still owns config.Config and will move in Phase 4.
