+++ app/go.mod
@@ -10,8 +10,11 @@ replace (
 require (
 	github.com/Ivantseng123/agentdock v0.0.0
 	github.com/Ivantseng123/agentdock/shared v0.0.0
+	github.com/fsnotify/fsnotify v1.9.0
 	github.com/slack-go/slack v0.20.0
 	github.com/xuri/excelize/v2 v2.10.1
+	golang.org/x/sync v0.20.0
+	gopkg.in/yaml.v3 v3.0.1
 )
 
 require (
@@ -40,5 +43,4 @@ require (
 	golang.org/x/sys v0.43.0 // indirect
 	golang.org/x/text v0.34.0 // indirect
 	google.golang.org/protobuf v1.36.8 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
