+++ app/skill/loader_test.go
