+++ app/skill/npx.go
