+++ app/skill/npx_test.go
