+++ app/skill/validate.go
