+++ app/skill/validate_test.go
